+++ packages/expo-router/src/matchers.tsx
@@ -64,3 +64,12 @@ export function stripGroupSegmentsFromPath(path: string): string {
 export function stripInvisibleSegmentsFromPath(path: string): string {
   return stripGroupSegmentsFromPath(path).replace(/\/?index$/, '');
 }
+
+/**
+ * Match:
+ *  - _layout files, +html, +not-found, string+api, etc
+ *  - Routes can still use `+`, but it cannot be in the last segment.
+ */
+export function isTypedRoute(name: string) {
+  return !name.startsWith('+') && name.match(/(_layout|[^/]*?\+[^/]*?)\.[tj]sx?$/) === null;
+}